feat: check if there r no users on past of redis

+++ src/services/update-sessionid.ts
@@ -21,8 +21,9 @@ export class UpdateSessionIdService{
         sessionId
       }
     })
-
-    await client.del(oldStudent.sessionId)
+    if(await client.hGetAll(oldStudent.sessionId))
+      await client.del(oldStudent.sessionId)
+    
     await client.hSet(sessionId, {
       username: updatedStudent.username,
       email: updatedStudent.email,
